pyproject.toml
[tool.poetry]
name = "opensqli"
version = "0.1.2"
description = "OpenAPI SQLi Scanner"
authors = ["Sergey M <tz4678@gmail.com>"]
repository = "https://github.com/tz4678/openapi-sqli-scanner"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
aiodns = "^3.0.0"
cchardet = "^2.1.7"
PyYAML = "^6.0"
aiolimiter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyterlab = "^3.3.2"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
opensqli = "opensqli.cli:main"
